requirements.txt
amqp==5.0.6
anyjson==0.3.3
asgiref==3.4.1
Babel==2.9.1
billiard==3.6.4.0
celery~=5.2.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
Django~=3.2.9
django-debug-toolbar~=3.2.2
flower==1.0.0
gunicorn==20.1.0
humanize==3.12.0
kombu==5.2.1
prometheus-client==0.8.0
prompt-toolkit==3.0.22
psycopg==3.0.3
psycopg-binary==3.0.3
psycopg-pool==3.0.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pytz==2021.3
redis==4.0.0
six==1.16.0
sqlparse==0.4.2
tornado==5.1.1
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.13.3

djangorestframework~=3.12.4
ipinfo~=4.2.1
pyowm~=3.2.0
peewee~=3.14.8